lyrikist/requirements.txt
   Flask==2.0.3
   Werkzeug==2.0.3
   python-dotenv==0.19.2
   requests==2.27.1
   deep-translator==1.10.1
   openai==0.10.2  # Add this line

   